% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_locations_into_ssdb.R
\name{add_locations_into_ssdb}
\alias{add_locations_into_ssdb}
\title{Add locations ("locations" table) to a SumavaDB-like spatiotemporal database}
\usage{
add_locations_into_ssdb(
  conn,
  locations,
  new = FALSE,
  append = !new,
  sql_files = system.file("sql/create_or_replace_functions.sql", package = "SuSnowDB"),
  warn = TRUE,
  err = !warn,
  nn_locations = c("location_code", "location_code0", "location_name", "altitude",
    "city_name", "country_code_iso_3166_1", "country_code_iso_3166_2", "country_name",
    "description", "geometry", "location_source", "location_url", "use_limitations"),
  ...
)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{locations}{locatiol, spatial object (or \code{\link{sf}}) . See output of \code{\link{sumava_snow_dataset}}}

\item{new}{logical. If \code{new==TRUE} database tables are created or replaced. If \code{new==FALSE} tables are modidied according to \code{append}. Default is \code{FALSE}.}

\item{append}{logical. If \code{append==FALSE} database tables are created or replaced. If \code{append==FALSE} data and metedata are appended as new rows of the database tables. Default is \code{append=TRUE}.}

\item{sql_files}{file names containing database initialization SQL queries. They are used in case \code{append==FALSE} or \code{new=TRUE}.}

\item{warn}{logical argument if the column names of \code{locations} are different from \code{nn_locations} functions displays a warning message. Default is \code{TRUE}.}

\item{err}{logical argument if the column names of \code{locations} are different from \code{nn_locations} functions returns an error. Default is \code{!warn}.}

\item{nn_locations}{names of the columns in the "locations" table. See the vignette.}

\item{...}{further arguments (currently not used)}
}
\description{
Add locations ("locations" table) to a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)
add_locations_into_ssdb(conn,val$locations,new=TRUE)
### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- st_read(conn,"locations")
dbDisconnect(conn)

}
}
