% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_measurement_types_into_ssdb.R
\name{add_measurement_types_into_ssdb}
\alias{add_measurement_types_into_ssdb}
\title{Add measurement_types ("measurement_types" table) to a SumavaDB-like spatiotemporal database}
\usage{
add_measurement_types_into_ssdb(
  conn,
  measurement_types,
  new = FALSE,
  append = !new,
  sql_files = system.file("sql/create_or_replace_functions.sql", package = "SuSnowDB"),
  warn = TRUE,
  err = !warn,
  nn_measurement_types = default_nn_measurement_types(),
  ...
)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{measurement_types}{measurement types table, data frame or an object that can be coerced to a data frame.. See output of \code{\link{sumava_snow_dataset}}}

\item{new}{logical. If \code{new==TRUE} database tables are created or replaced. If \code{new==FALSE} tables are modidied according to \code{append}. Default is \code{FALSE}.}

\item{append}{logical. If \code{append==FALSE} database tables are created or replaced. If \code{append==FALSE} data and metedata are appended as new rows of the database tables. Default is \code{append=TRUE}.}

\item{sql_files}{file names containing database initialization SQL queries. They are used in case \code{append==FALSE} or \code{new=TRUE}.}

\item{warn}{logical argument if the column names of \code{measurement_types} are different from \code{nn_measurement_types} functions displays a warning message. Default is \code{TRUE}.}

\item{err}{logical argument if the column names of \code{measurement_types} are different from \code{nn_measurement_types} functions returns an error. Default is \code{!warn}.}

\item{nn_measurement_types}{names of the columns in the "measurement_types" table. See the vignette.}

\item{...}{further arguments (currently not used)}
}
\description{
Add measurement_types ("measurement_types" table) to a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)
add_measurement_types_into_ssdb(conn,val$measurement_types,new=TRUE)
### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_measurement_types_into_ssdb(conn,val_smet$measurement_types,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- dbReadTable(conn,"measurement_types")
dbDisconnect(conn)

}
}
