% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_measurements_into_ssdb.R
\name{add_measurements_into_ssdb}
\alias{add_measurements_into_ssdb}
\title{Add measurements ("measurements" table) to a SumavaDB-like spatiotemporal database}
\usage{
add_measurements_into_ssdb(
  conn,
  measurements,
  new = FALSE,
  append = !new,
  sql_files = system.file("sql/create_or_replace_functions.sql", package = "SuSnowDB"),
  warn = TRUE,
  err = !warn,
  nn_measurements = c("time", "value", "flag", "location_code0", "variable_code0",
    "description"),
  ...
)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{measurements}{measurements table, data frame or an object that can be coerced to a data frame. See output of \code{\link{sumava_snow_dataset}}}

\item{new}{logical. If \code{new==TRUE} database tables are created or replaced. If \code{new==FALSE} tables are modidied according to \code{append}. Default is \code{FALSE}.}

\item{append}{logical. If \code{append==FALSE} database tables are created or replaced. If \code{append==FALSE} data and metedata are appended as new rows of the database tables. Default is \code{append=TRUE}.}

\item{sql_files}{file names containing database initialization SQL queries. They are used in case \code{append==FALSE} or \code{new=TRUE}.}

\item{warn}{logical argument if the column names of \code{measurements} are different from \code{nn_measurements} functions displays a warning message. Default is \code{TRUE}.}

\item{err}{logical argument if the column names of \code{measurements} are different from \code{nn_measurements} functions returns an error. Default is \code{!warn}.}

\item{nn_measurements}{names of the columns in the "measurements" table. See the vignette.}

\item{...}{further arguments (currently not used)}
}
\description{
Add measurements ("measurements" table) to a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)

add_locations_into_ssdb(conn,val$locations,new=TRUE)
add_measurement_types_into_ssdb(conn,val$measurement_types)
add_measurements_into_ssdb(conn,val$measurements)


### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,append=TRUE)
add_measurement_type_into_ssdb(conn,val_smet$measurement_type,append=TRUE)
add_measurements_into_ssdb(conn,val_smet$measurements,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- dbReadTable(conn,"measurements")
dbDisconnect(conn)

}
}
