% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_dataset_from_ssdb.R
\name{crop_dataset_from_ssdb}
\alias{crop_dataset_from_ssdb}
\title{Crop or get a dataset from ssdb}
\usage{
crop_dataset_from_ssdb(
  ssdb,
  crop = NULL,
  table_name = get_value_table_names_from_ssdb(ssdb),
  variables_code0 = NULL,
  ...
)
}
\arguments{
\item{ssdb}{database (preferably PostgresSQL) connection (see output of \code{\link{dbConnect}}) or on  a generic object or a dataset, e.g. an object returned by \code{\link{gsodr_dataset}} OR a list of these objects.}

\item{crop}{argument \code{"y"} of \code{\link{st_crop}}}

\item{table_name}{vector with names of the tables . Default is \code{get_value_table_names_from_ssdb(conn)}. See also @seealso \code{\link{get_value_table_names_from_ssdb}}.}

\item{variables_code0}{code of the types (\code{variable_code0}) of the measurements that are requested to be extracted.}

\item{...}{further arguments for  \code{\link{get_measurements_from_ssdb}}}
}
\description{
Crop or get a dataset from ssdb
}
\examples{

library(magrittr)
library(raster)

ssdb <- '/home/ecor/activity/2022/local/shiny/SuSnowApp/ext-data/sumava.rds' \%>\% readRDS()

crop_ext <- '/home/ecor/activity/2021/local/SuSnowDB/inst/sumava_extdata/sumava_v2_extent.tif' \%>\% raster() 

ssdb_cropped <- crop_dataset_from_ssdb(ssdb,crop=crop_ext)


}
