% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_union.R
\name{datasets_union}
\alias{datasets_union}
\title{Union of one or more Weather Datasets from (UDER DEVELOPMENT)}
\usage{
datasets_union(
  x,
  ...,
  warn = FALSE,
  err = !warn,
  nn_measurement_types = default_nn_measurement_types(),
  nn_locations = default_nn_locations(),
  nn_measurements = default_nn_measurements()
)
}
\arguments{
\item{x}{dataset, e.g. an object returned by \code{\link{gsodr_dataset}}}

\item{...}{further arguments, e.g other object like \code{x}.}

\item{warn}{logical argument if any elements of \code{x,...} are not dataset-like objects, function displays a warning message. Default is \code{TRUE}.}

\item{err}{logical argument if any elements of \code{x,...} are not dataset-like objects, function returns an error. Default is \code{!warn}.}

\item{nn_measurement_types}{see \code{\link{add_measurement_types_into_ssdb}}.}

\item{nn_locations}{see \code{\link{add_locations_into_ssdb}}.}

\item{nn_measurements}{see \code{\link{add_measurements_into_ssdb}}.}
}
\description{
Union of one or more Weather Datasets from (UDER DEVELOPMENT)
}
\examples{
 library(readr)
 x <- system.file("west_africa_gsod_extdata/gsod_bf_v2.csv",package="SuSnowDB") 
 x <- x \%>\% read_csv() \%>\% gsodr_dataset()
 y <- system.file("west_africa_gsod_extdata/gsod_ml.csv",package="SuSnowDB") 
 y <- y \%>\% read_csv() \%>\% gsodr_dataset()
 z <- system.file("west_africa_gsod_extdata/gsod_sn.csv",package="SuSnowDB") 
 z <- z \%>\% read_csv() \%>\% gsodr_dataset()

 out <- datasets_union(x,y,z)
 
 outl <- datasets_union(list(x,y,z))
 
 is_dataset(out)
 is_dataset(outl)
 
}
