% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_nn.R
\name{default_nn_measurement_types}
\alias{default_nn_measurement_types}
\alias{default_nn_locations}
\alias{default_nn_measurements}
\title{Column names of the SuSnowDB Tables}
\usage{
default_nn_measurement_types(
  nn = c("variable_code0", "variable", "unit", "description",
    "measurement_time_interval")
)

default_nn_locations(
  nn = c("location_code", "location_code0", "location_name", "altitude", "city_name",
    "country_code_iso_3166_1", "country_code_iso_3166_2", "country_name", "description",
    "geometry", "location_source", "location_url", "use_limitations")
)

default_nn_measurements(
  nn = c("time", "value", "flag", "location_code0", "variable_code0", "description")
)
}
\arguments{
\item{nn}{names of the columns in the "locations" table. See the vignette.}
}
\description{
See Vignette
}
\examples{

default_nn_measurement_types()
default_nn_measurements()
default_nn_locations()


}
