% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations_from_ssdb.R
\name{get_locations_from_ssdb}
\alias{get_locations_from_ssdb}
\title{Get locations ("locations" table) from a SumavaDB-like spatiotemporal database}
\usage{
get_locations_from_ssdb(conn, ...)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{...}{further arguments for \code{\link{st_read}}}
}
\description{
Get locations ("locations" table) from a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)
add_locations_into_ssdb(conn,val$locations,new=TRUE)
### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- get_locations_from_ssdb(conn)
dbDisconnect(conn)

}
}
