% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurement_types_from_ssdb.R
\name{get_measurement_types_from_ssdb}
\alias{get_measurement_types_from_ssdb}
\title{Get variables ("measurement_types" table) from a SumavaDB-like spatiotemporal database}
\usage{
get_measurement_types_from_ssdb(conn, ...)
}
\arguments{
\item{conn}{database (preferably PostgresSQL) connection. See output of \code{\link{dbConnect}}.}

\item{...}{further arguments for \code{\link{dbReadTable}}}
}
\description{
Get variables ("measurement_types" table) from a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
val <- sumava_snow_dataset()
conn = dbConnect(PostgreSQL(), dbname = dbname)
add_measurement_types_into_ssdb(conn,val$measurement_types,new=TRUE)
### Add weather station from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_measurement_types_into_ssdb(conn,val_smet$measurement_types,append=TRUE)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- get_measurement_types_from_ssdb(conn)
dbDisconnect(conn)

}
}
