% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_table_names_from_ssdb.R
\name{get_value_table_names_from_ssdb}
\alias{get_value_table_names_from_ssdb}
\title{Extract table names containing values}
\usage{
get_value_table_names_from_ssdb(
  ssdb,
  meta_table_names = c("locations", "measurement_types"),
  value_table_names = NULL
)
}
\arguments{
\item{ssdb}{spatio-temporal dataset or a connection (TO BE IMPLEMENTED)}

\item{meta_table_names}{names of the meta tables. Default is \code{c("locations","measurement_types")}.}

\item{value_table_names}{names of the meta table. Default is \code{NULL}. Ignored if it is \code{NULL}.}
}
\description{
Extract table names containing values
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
conn = dbConnect(PostgreSQL(), dbname = dbname)


### Add weather data from stations from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,new=TRUE)
add_measurement_types_into_ssdb(conn,val_smet$measurement_type)
add_measurements_into_ssdb(conn,val_smet$measurements)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- get_measurements_from_ssdb(conn)
out <- get_measurements_from_ssdb(conn,locations_code0="IT-TN_T0179")
out <- get_measurements_from_ssdb(conn,locations_code0="IT-TN_T0179",variables_code0=c("SMET_TA","SMET_PINT"))
out <- get_measurements_from_ssdb(conn,variables_code0=c("SMET_TA","SMET_PINT"))

time_interval <- as.Date(c("2021-03-01","2021-03-31"))
out <- get_measurements_from_ssdb(conn,variables_code0=c("SMET_TA","SMET_PINT"),start_time=time_interval[1],end_time=time_interval[2])

tnames <- get_value_table_names_from_ssdb(conn)
dbDisconnect(conn)


sumava <- sumava_snow_dataset()
tnames_sumava <- get_value_table_names_from_ssdb(sumava)
}
}
