% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsodr_defaults.R
\name{gsod_defaults}
\alias{gsod_defaults}
\title{GSOD defaults}
\usage{
gsod_defaults(nn = NULL)
}
\arguments{
\item{nn}{argument}
}
\description{
GSOD defaults
}
\examples{


defaults <- gsod_defaults()

}
