% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsodr_dataset.R
\name{gsodr_dataset}
\alias{gsodr_dataset}
\title{Create GSOD Weather Dataset from (UDER DEVELOPMENT)}
\usage{
gsodr_dataset(
  x = NULL,
  ...,
  description_url = "https://www7.ncdc.noaa.gov/CDO/GSOD_DESC.txt",
  location_code0 = "GSOD_\%s",
  variable_code0 = "GSODR_\%s",
  location_url = gsod_defaults("location_url"),
  use_limitations = gsod_defaults("use_limitations"),
  location_source = gsod_defaults("location_source"),
  crs = 4326,
  measurement_types = gsod_defaults("measurement_types")
)
}
\arguments{
\item{x}{data frame returned from a call of \code{\link{get_GSOD}}. Default is \code{NULL}.}

\item{...}{arguments for \code{\link{get_GSOD}}}

\item{description_url, location_url, location_source, use_limitations}{attributes for \code{locations} table. See \code{locations} table and default.}

\item{location_code0, variable_code0}{attributes for \code{locations} and \code{measurement_types} tables, as above. If it is one-element,it may contain a sub-string "\code{\%s}" that will be replaced with the GSOD station identification code (STNID).}

\item{crs}{Coordinate refence system. Default is 4326 (epsg).}

\item{measurement_types}{data frame from a CSV file containing table with measuremet type attributes . It must contain an additional column \code{has_flag} .See \code{measurement_types} table and default.}
}
\description{
Create GSOD Weather Dataset from (UDER DEVELOPMENT)
}
\examples{
 library(readr)
 x <- system.file("west_africa_gsod_extdata/gsod_bf.csv",package="SuSnowDB") \%>\% read_csv()
 out <- gsodr_dataset(x=x)
 
 \dontrun{
 t <- system.time(out <- gsodr_dataset(years=2020,country="IT"))
 }
}
