% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dataset.R
\name{is_dataset}
\alias{is_dataset}
\alias{are_all_datasets}
\title{Is it a dataset formatted for a SuSnowDB ? Are they all datasets formatted for a SuSnowDB ?}
\usage{
is_dataset(
  x,
  nn_measurement_types = default_nn_measurement_types(),
  nn_locations = default_nn_locations(),
  nn_measurements = default_nn_measurements()
)

are_all_datasets(
  x,
  ...,
  nn_measurement_types = default_nn_measurement_types(),
  nn_locations = default_nn_locations(),
  nn_measurements = default_nn_measurements()
)
}
\arguments{
\item{x}{a generic object or a dataset, e.g. an object returned by \code{\link{gsodr_dataset}} OR a list of these objects.}

\item{nn_measurement_types}{see \code{\link{add_measurement_types_into_ssdb}}.}

\item{nn_locations}{see \code{\link{add_locations_into_ssdb}}.}

\item{nn_measurements}{see see \code{\link{add_measurements_into_ssdb}}.}

\item{...}{further arguments, other object like \code{x}}
}
\description{
Is it a dataset formatted for a SuSnowDB ? Are they all datasets formatted for a SuSnowDB ?
}
\examples{
 library(readr)
 x <- system.file("west_africa_gsod_extdata/gsod_bf.csv",package="SuSnowDB") \%>\% read_csv()
 val <- gsodr_dataset(x=x)
 out <- is_dataset(val)
 
 val2 <- data.frame(a=3,b=1:10)
 out2 <- is_dataset(val2)

 are_all_datasets(val,val2)
 are_all_datasets(list(val,val2))
 
}
