% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteotrentino_defaults.R
\name{meteotrentino_defaults}
\alias{meteotrentino_defaults}
\title{Meteotrentino defaults}
\usage{
meteotrentino_defaults(nn = NULL)
}
\arguments{
\item{nn}{argument}
}
\description{
Meteotrentino defaults
}
\examples{


defaults <- meteotrentino_defaults()

}
