% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumava_snow_dataset.R
\name{sumava_snow_dataset}
\alias{sumava_snow_dataset}
\title{Create Sumava Snow Dataset from DWD and CHMI sources}
\usage{
sumava_snow_dataset(
  data_dir = system.file("snow_extdata", package = "SuSnowDB"),
  remove_multipoints = TRUE
)
}
\arguments{
\item{data_dir}{directory containing original CHMI and DWD snow data}

\item{remove_multipoints}{logical. It splits station and related time series if a station has resulted in different (even if near) point sites in different time periods (e.g. some CHMI stations). DEfault is \code{TURE}.}
}
\description{
Create Sumava Snow Dataset from DWD and CHMI sources
}
\examples{
 data_dir <- system.file("snow_extdata",package="SuSnowDB")
 \dontrun{
 out <- sumava_snow_dataset()
 is_dataset(out)
 }
}
