% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablenames.R
\name{tablenames}
\alias{tablenames}
\alias{tablenames,list-method}
\alias{tsblenames}
\alias{tablenames,PostgreSQLConnection-method}
\title{Lorem ipsum}
\usage{
tablenames(ssdb = NULL, ...)

tablenames(ssdb = NULL, ...)

\S4method{tablenames}{list}(ssdb = NULL, ...)

\S4method{tablenames}{PostgreSQLConnection}(
  ssdb = NULL,
  ignored = c("_hyper_29_34918_chunk", "hypertable", "hypertable_data_node",
    "tablespace", "dimension", "dimension_slice", "chunk", "chunk_constraint",
    "chunk_index", "chunk_data_node", "metadata", "continuous_agg",
    "continuous_aggs_invalidation_threshold",
    "continuous_aggs_hypertable_invalidation_log",
    "continuous_aggs_materialization_invalidation_log", "hypertable_compression",
    "compression_algorithm", "compression_chunk_size", "remote_txn", "bgw_job",
    "bgw_job_stat", "bgw_policy_chunk_stats", "cache_inval_hypertable",     
    "cache_inval_bgw_job", "cache_inval_extension", "spatial_ref_sys", "hyper", "chunk"),
  all = FALSE,
  ...
)
}
\arguments{
\item{ssdb}{object}

\item{...}{further arguments}

\item{ignored}{ingored table names that are ignored in case \code{all==FALSE}.}

\item{all}{logical. If it is \code{TRUE} no names are ignored. Default is \code{FALSE}.}
}
\description{
Lorem ipsum
}
\details{
Lorem ipsum
}
\examples{
\dontrun{
library(RPostgreSQL) 
dbname <- "hydroclimatedb_testing00"
conn = dbConnect(PostgreSQL(), dbname = dbname)


### Add weather data from stations from MeteoTrentino network, Trentino, Italy 
val_smet <- meteotrentino_smet_dataset(smet_files=c("T0175","T0179"))
add_locations_into_ssdb(conn,val_smet$locations,new=TRUE)
add_measurement_types_into_ssdb(conn,val_smet$measurement_type)
add_measurements_into_ssdb(conn,val_smet$measurements)

dbDisconnect(conn)

conn = dbConnect(PostgreSQL(), dbname = dbname)
out <- get_measurements_from_ssdb(conn)
out <- get_measurements_from_ssdb(conn,locations_code0="IT-TN_T0179")
out <- get_measurements_from_ssdb(conn,locations_code0="IT-TN_T0179",variables_code0=c("SMET_TA","SMET_PINT"))
out <- get_measurements_from_ssdb(conn,variables_code0=c("SMET_TA","SMET_PINT"))

time_interval <- as.Date(c("2021-03-01","2021-03-31"))
out <- get_measurements_from_ssdb(conn,variables_code0=c("SMET_TA","SMET_PINT"),start_time=time_interval[1],end_time=time_interval[2])

tnames <- tablenames(conn)
dbDisconnect(conn)


sumava <- sumava_snow_dataset()
tnames_sumava <- tablenames(sumava)
}

}
