% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl,list-method}
\alias{tbl,list-method}
\alias{tbl}
\title{tbl}
\usage{
\S4method{tbl}{list}(src, ...)
}
\arguments{
\item{src}{argument}

\item{...}{argument}
}
\description{
"tbl" implemantation for a list
}
