% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_append.R
\name{write_dataset_append}
\alias{write_dataset_append}
\title{Append "dataset" to a SmavaDB-like spatiotemporal database}
\usage{
write_dataset_append(conn, dataset, append = TRUE, ...)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{dataset}{dataset . See output of \code{\link{write_dataset_append}}}

\item{append, ...}{further arguments. Default is \code{append=TRUE}.}
}
\description{
Append "dataset" to a SmavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 

val <- sumava_snow_dataset()
dbname = "hydroclimatedb_ver03"
conn = dbConnect(PostgreSQL(), dbname = dbname)
write_dataset_append(conn,dataset)
dbDisconnect(conn) 
}
}
