% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset_into_ssdb.R
\name{write_dataset_into_ssdb}
\alias{write_dataset_into_ssdb}
\title{Append "dataset" to a SumavaDB-like spatiotemporal database}
\usage{
write_dataset_into_ssdb(
  conn,
  dataset,
  append = TRUE,
  new = FALSE,
  sql_files = system.file("sql/create_or_replace_functions.sql", package = "SuSnowDB"),
  ...
)
}
\arguments{
\item{conn}{database (preferrably PodtgresSQL) connection. See ouput of \code{\link{dbConnect}}.}

\item{dataset}{dataset . See output of \code{\link{sumava_snow_dataset}}}

\item{append}{logical. If \code{append==FALSE} database tables are created or replaced. If \code{append==FALSE} data and metedata are appended as new rows of the database tables. Default is \code{append=TRUE}.}

\item{new}{logical. If \code{new==TRUE} database tables are created or replaced. If \code{new==FALSE} tables are modidied according to \code{append}. Default is \code{FALSE}.}

\item{sql_files}{file names containing database initialization SQL queries. They are used in case \code{append==FALSE} or \code{new=TRUE}.}

\item{...}{further arguments.}
}
\description{
Append "dataset" to a SumavaDB-like spatiotemporal database
}
\examples{
\dontrun{
library(RPostgreSQL) 

val <- sumava_snow_dataset()
dbname = "hydroclimatedb_ver03"
conn = dbConnect(PostgreSQL(), dbname = dbname)

## Create a new DB layout
write_dataset_into_ssdb(conn,dataset=val,new=TRUE)
dbDisconnect(conn) 
}
}
