% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumava_snow_dataset.R
\name{sumava_snow_dataset}
\alias{sumava_snow_dataset}
\title{Create Sumava Snow Dataset from DWD and CHMI sources}
\usage{
sumava_snow_dataset(
  data_dir = system.file("snow_extdata", package = "SuSnowDB")
)
}
\arguments{
\item{data_dir}{directory containing original CHMI and DWD snow data}
}
\description{
Create Sumava Snow Dataset from DWD and CHMI sources
}
\examples{
 data_dir <- system.file("snow_extdata",package="SuSnowDB")
 \dontrun{
 out <- sumava_snow_dataset()
 }
}
